cleanup: Use the Rust memory allocator directly.

+++ util.c
@@ -353,13 +353,12 @@ notincl (char *feature)
 }
 
 /* grow a static string to at least a certain length */
-
 void
 growstr (char **strptr, int *curlen, int newlen)
 {
     if (newlen > *curlen) {		/* need more room? */
 	if (*curlen)
-	    *strptr = saferealloc(*strptr,(size_t)newlen);
+	    *strptr = saferealloc(*strptr,newlen);
 	else
 	    *strptr = safemalloc(newlen);
 	*curlen = newlen;
